Save original system configuration files before changing them
The scripts LCD35-show and LCD-hdmi change configuration files to
setup everything to activate or deactivate the LCD screen respectively.
With this commit, those scripts will backup the files to be changed
before doing so when switching to LCD display, and restore them back
from the backup when switching back to HDMI display.